requirements.txt
certifi==2022.6.15.1
discord.py==2.0.1
python-dotenv==0.21.0
redis==4.5.1
requests==2.28.1
SQLAlchemy==2.0.23
